requirements.txt
click
google-api-core>=2.11.0
types-requests
